list

+++ src/main/java/com/spring/repository/MarqueRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.spring.models.Marque;
 
-public interface MarqueRepository extends JpaRepository<Marque, Long> {
+public interface MarqueRepository extends JpaRepository<Marque, Integer> {
 
 }
